requirements.txt
drequests==2.25.1
beautifulsoup4==4.9.3
python_dateutil==2.8.2
